==> R/Part_2_Description_sample_table1.Rmd <==
---
title: "Preoperative Atelectasis"
subtitle: "Part 2: Descriptive characteristics of the sample and Table 1"
author: "Javier Mancilla Galindo"
date: last rendered on "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    theme: default
    highlight: pygments
---

# Setup 
```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE) # show both code and output by default
knitr::opts_chunk$set(warning = FALSE) # hides all warnings
knitr::opts_chunk$set(message = FALSE) # hides all package message
```

#### Packages used   
```{r}
if (!require("pacman", quietly = TRUE)) {
  install.packages("pacman")
}

pacman::p_load(
  tidyverse, # Used for basic data handling and visualization.
  table1, #Used to create table of descriptive characteristics of sample.
  RColorBrewer, #Color palettes for data visualization. 
  gridExtra, #Used to arrange multiple ggplots in a grid.  
  grid, #Used to arrange multiple ggplots in a grid.
  rnaturalearth, #Used to extract geographical data to create maps. 
  sf, #Used together with the prior package to create map. 
  plotly, #Used together with prior two packages to create map.  
  flextable, #Used to export tables.  
  officer  #Used to export tables.
)
```

##### Session   
```{r, echo=FALSE}
# Credits chunk of code: Alex Bossers, Utrecht University (a.bossers@uu.nl)

# remove clutter
session <- sessionInfo()
session$BLAS <- NULL
session$LAPACK <- NULL
session$loadedOnly <- NULL
# write log file
writeLines(
  capture.output(print(session, locale = FALSE)),
  paste0(lubridate::today(), "_session_Part_2.txt")
)

session
```

```{r, include = FALSE}
rm(session)
```

```{r, include = FALSE}
# Create directories for sub-folders 
figfolder <- "../results/01_output_figures"
tabfolder <- "../results/01_output_tables"
psfolder <- "../data/processed"
dir.create(figfolder, showWarnings = FALSE)
dir.create(tabfolder, showWarnings = FALSE)
dir.create(psfolder, showWarnings = FALSE)
```


```{r, include = FALSE}
# Load dataset  
data <- read.csv("../data/processed/atelectasis_included.csv",
                 na.strings="NA", 
                 row.names = NULL)
```

```{r, include = FALSE}
# Recode variables 
source("variable_names_attributes.R", local = knitr::knit_global())
```

# State of residence of participants  

Map generated with the accompanying script ***Map_USA_Canada.R***

This map was built by partly using code adapted from [contribution by cpsievert](https://community.rstudio.com/t/combining-usa-canada-specific-province-with-plotly-plot-geo-or-ggplot2/12375).

```{r, echo=FALSE}
source("Map_USA_Canada.R", local = knitr::knit_global())
fig
```

```{r, include = FALSE}
rm(canada, usa, fig, fig_add, figS2, table, us_laea)
```


# Distribution of numerical variables    

Distributions were examined with the accompanying sourced function      ***distribution_numerical_variables.R***      
```{r, include=FALSE, fig.height = 4, fig.width = 6}
source("distribution_numerical_variables.R", local = knitr::knit_global())
distribution_plots(data)
```

Near normal distribution:  
- Age: light tails  
- height: heavy right tail, 4 outliers right  
- hb: heavy tails, bilateral outliers    
- hct: heavy tails, bilateral outliers  
- leu: near normal, bilateral outliers  
- neu_absolute: heavy right tail, two right outliers  
- linf_absolute: heavy right tail, bilateral outliers (more right)     
- mon_absolute: heavy right tail, bilateral outliers (more right)   
- platelets: two right outliers   
- urea: four right outliers  
- creatinine: three right outliers  


Distribution not normal:  
- Weight: right-skewed, outliers are verified observations of extreme weight.   
- BMI: right-skewed, outliers are verified observations of extreme BMI.  
- spo2_VPO: Left-skewed   
- neu_percent: left-skewed  
- linf_percent: right-skewed  
- glucose: right-skewed  
- mon_percent: observations around only 5 data points. Will not use this variable, only absolute monocytes will be used.     
- altitude: distribution not clear as values are quite apart an concentrate around single states with diferring mean altitudes. Will attempt to model a smooth term or categorical term in subsequent analyses.         

Outcome variable:  
- atelectasis_percent: Zero-inflated. Would be difficult to manage as categorical ordinal due to low number of patients in some categories. Will re-assess alongside subsequent analyses to decide.   


# Table 1   
Table 1 generated with the accompanying sourced script ***table1_arguments.R***    
```{r, include=FALSE}
source("table1_arguments.R", local = knitr::knit_global())
```

Characteristics of participants are shown for the total sample and by obesity class category as defined by the [World Health Organization](https://www.who.int/europe/news-room/fact-sheets/item/a-healthy-lifestyle---who-recommendations):   

- Class 1, BMI [30,35) kg/m2    
- Class 2, BMI [35,40) kg/m2     
- Class 3, BMI >40 kg/m2     

```{r, echo=FALSE}
table1
```

```{r, include = FALSE}
rm(properties, name, table1, rndr, normal, nonormal, abbreviations, abbreviations_stats)
```


NOTE: The **ASA** physical status variable has not been included in analyses since the [updated version of ASA](https://www.asahq.org/standards-and-practice-parameters/statement-on-asa-physical-status-classification-system) consulted in October 2023 classifies obesity (30<BMI<40) as ASA 2 and obesity (BMI ≥40) as ASA 3. The distribution of frequencies of ASA~obesity class in this dataset does not match such definition. This occurred since an outdated version of ASA that did not include obesity was likely used by clinicians when writing the preoperative assessment medical note:   

```{r, echo=FALSE}
table(data$ASA,data$type_obesity)
```

```{r, include=FALSE}
rm(data, figfolder, psfolder, tabfolder)
pacman::p_unload(negate = TRUE)
```
